requirements.txt
git+https://github.com/openai/whisper.git
openai
langchain
youtube_search
pandas
yt-dlp
whisper
pytube
python-dotenv
youtube-transcript-api
chromadb
google-search-results
wikipedia
streamlit
